+++ R/export.r
@@ -8,14 +8,10 @@
 #' in format used by \pkg{OpenRepGrid}. This file format can also 
 #' easily be edited by hand (see \code{\link{importTxt}} for a 
 #' description).
-#' The function will open an interactive dialog box to let the user 
-#' enter a filename if no \code{file} argument is supplied
-#' in the function call. 
 #'
 #' @param x     \code{repgrid} object.
 #' @param file  Filename to save the grid to. The name should have 
-#'              the suffix .txt. If the function is called without
-#'              specifying this argument a dialog box is opened.
+#'              the suffix \code{.txt}.
 #' @return      Invisibly returns the name of the file.
 #'
 #' @note
@@ -61,13 +57,8 @@
 #'
 #' }
 #'
-saveAsTxt <- function(x, file=NA){  
-	if (is.na(file)){
-	  #require(tcltk)
-  	fileName <- tclvalue(tkgetSaveFile())     # open filename dialog  
-	} else 
-	  fileName <- file
-	
+saveAsTxt <- function(x, file = NA) {  
+  fileName <- file
 	enames <- elements(x)
 	cnames <- constructs(x)
 	scores <- getRatingLayer(x)